Renaming and reorganizing lib files into packages with descriptive names
