+++ pyproject.toml
@@ -36,6 +36,8 @@ dependencies = [
     "numpy>=1.19.5,<1.24.0",
     "pandas>=1.1.5,<1.5.0",
     "jsonlines>=3.1.0,<3.2.0",
+    "openai",
+    "ndjson"
 ]
 
 [project.license]
@@ -80,7 +82,7 @@ content-type = "text/markdown"
 
 
 [project.entry-points.spacy_factories]
-"conspiracies/prompt_relation_extraction" = "conspiracies.prompt_relation_extraction.prompt_relation_component:create_prompt_relation_extraction_component"
+"conspiracies/relationextraction/gptprompting" = "conspiracies.relationextraction.gptprompting:create_prompt_relation_extraction_component"
 
 
 [build-system]
