+++ src/conspiracies/__init__.py
@@ -1,16 +1,2 @@
-from .HeadWordExtractionComponent import HeadwordsExtractionComponent  # noqa F401
-from .HeadWordExtractionComponent import create_headwords_component  # noqa F401
 from .registry import registry  # noqa F401
-from .prompt_relation_extraction import (  # noqa F401
-    PromptTemplate1,
-    PromptTemplate2,
-    MarkdownPromptTemplate1,
-    MarkdownPromptTemplate2,
-    XMLStylePromptTemplate,
-    chatGPTPromptTemplate,
-    SpanTriplet,
-    StringTriplet,
-    DocTriplets,
-)
-from .utils import docs_from_jsonl, docs_to_jsonl  # noqa F401
-from .wordpiece_length_normalization import wordpiece_length_normalization  # noqa F401
+from .doc_utils import docs_from_jsonl, docs_to_jsonl  # noqa F401
