+++ src/conspiracies/coref/__init__.py
@@ -1,2 +1,2 @@
-from .CoreferenceModel import CoreferenceModel  # noqa
-from .CoreferenceComponent import CoreferenceComponent, create_coref_component  # noqa
+from .coref_model import CoreferenceModel  # noqa
+from .coref_comp import CoreferenceComponent, create_coref_component  # noqa
