+++ src/conspiracies/relationextraction/__init__.py
@@ -1,2 +1,2 @@
-from .knowledge_triplets import KnowledgeTriplets  # noqa F401
-from .wrap_model_spacy import SpacyRelationExtractor  # noqa F401
+from .multioie2.knowledge_triplets import KnowledgeTriplets  # noqa F401
+from .multioie2.multi2oie_comp import SpacyRelationExtractor  # noqa F401
