+++ src/conspiracies/relationextraction/gptprompting/data_classes.py
