+++ src/conspiracies/relationextraction/gptprompting/prompt_apis.py
@@ -2,10 +2,10 @@ import random
 import time
 from typing import Any, Dict, List
 
-from ..registry import registry
+from conspiracies.registry import registry
 from spacy.tokens import Doc
 
-from conspiracies.prompt_relation_extraction import PromptTemplate
+from conspiracies.relationextraction.gptprompting.prompt_templates import PromptTemplate
 
 
 @registry.prompt_apis.register("conspiracies/openai_gpt3_api")
@@ -77,8 +77,9 @@ def create_openai_chatgpt_prompt_api(
 
         openai.api_key = api_key
         message_example = prompt_template.create_prompt("test")
-        assert (
-            type(message_example) == list and type(message_example[0]) == dict
+        assert isinstance(message_example, list) and isinstance(
+            message_example[0],
+            dict,
         ), "ChatGPT requires a list of message dicts. Consider using chatGPTPromptTemplate as template."  # noqa: E501
 
         responses: List[str] = []
