+++ src/conspiracies/relationextraction/gptprompting/prompt_templates.py
