+++ src/conspiracies/relationextraction/multioie2/__init__.py
@@ -1,0 +1,2 @@
+from .knowledge_triplets import KnowledgeTriplets  # noqa F401
+from .multi2oie_comp import SpacyRelationExtractor  # noqa F401
