+++ tests/test_utils.py
@@ -2,7 +2,8 @@ from pathlib import Path
 
 import pytest
 import spacy
-from conspiracies import DocTriplets, SpanTriplet, docs_from_jsonl, docs_to_jsonl
+from conspiracies import docs_from_jsonl, docs_to_jsonl
+from conspiracies.relationextraction.gptprompting import DocTriplets, SpanTriplet
 from spacy.tokens import Doc
 
 from .utils import docs_with_triplets  # noqa: F401
